autofutures-rt-v2/autofutures-rt/requirements.txt
python-binance==1.0.19
websockets==12.0
pandas==2.2.2
pandas-ta==0.3.14b
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
